% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_helpers.R
\name{check_chart}
\alias{check_chart}
\alias{check_chart_aspect_ratio}
\title{Check that your chart looks alright in the different chart formats}
\usage{
check_chart(type = "normal", object = ggplot2::last_plot())

check_chart_aspect_ratio(object = ggplot2::last_plot(), type = "normal")
}
\arguments{
\item{type}{The chart type (e.g. \code{"normal"}, \code{"fullslide"}).}

\item{object}{A ggplot2 plot object. Defaults to the last plot.}
}
\description{
Opens your plot in a new window with the aspect ratio of the chart type
selected. For fullslide chart types (fullslide, fullslide_narrow,
fullslide_half), this will show the complete slide with the grey header,
CIE logo, and properly positioned title/subtitle/caption.
}
\examples{
\dontrun{
ggplot(mtcars, aes(x = wt, y = mpg)) + geom_point()
check_chart("normal")

# Check fullslide appearance with title and subtitle
ggplot(mtcars, aes(x = wt, y = mpg)) +
  geom_point() +
  labs(title = "My title", subtitle = "My subtitle") +
  theme_cie()
check_chart("fullslide")
}

}
