% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cie_colours.R
\docType{data}
\name{cie_grey_title}
\alias{cie_grey_title}
\title{cie_grey_title}
\format{
An object of class \code{character} of length 1.
}
\usage{
cie_grey_title
}
\description{
cie_grey_title
}
\keyword{datasets}
