% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cie_label.R
\name{cie_label_functions}
\alias{cie_label_functions}
\alias{cie_label}
\alias{cie_label_repel}
\title{Create labels in the CIE style - no border, minimal padding}
\usage{
cie_label(
  ...,
  size = 18,
  padding = 0.1,
  lineheight = 0.8,
  linewidth = NA,
  fill = "white"
)

cie_label_repel(
  ...,
  size = 18,
  padding = 0.1,
  lineheight = 0.8,
  label.size = NA,
  fill = "white"
)
}
\arguments{
\item{...}{Arguments passed on to either `ggplot2::geom_label()` (in the
case of `cie_label()`) or `ggrepel::geom_label_repel()` (in the case of
`cie_label_repel()`). See \code{?geom_label} or \code{?geom_label_repel}.}

\item{size}{Font size for label. Default is 18.}

\item{padding}{Amount of white padding around label, measured in "lines"
(see \code{?unit::grid}). Default is 0.1.}

\item{lineheight}{Height of lines of text - smaller means the lines
of text are closer together. Default is `0.8`.}

\item{linewidth}{Size of label border line for `cie_label`, in mm. Default is `NA`, which means no border.}

\item{fill}{Colour of label background; default is `"white"`.}

\item{label.size}{Size of label border line for `cie_label_repel`, in mm. Default is `NA`, which means no border.}
}
\description{
`cie_label()` and `cie_label_repel()` create labels using
`ggplot2::geom_label()` and `ggrepel::geom_label_repel()` respectively. The
labels will have no border and minimal padding, in keeping with The CIE visual style.
}
\details{
`cie_label()` labels, just like `ggplot::geom_label()` labels, will be places at the
x, y coordinates given by the data. `cie_label_repel()` labels, just like
`ggrepel::geom_label_repel()` labels, will be placed near their x, y
coordinates, but will be places so that the labels don't overlap each other.
}
\examples{
library(ggplot2)

# Create a dataset -- mtcars with rownames as column
mt <- mtcars
mt$car <- rownames(mt)

# Create a plot without labels
p <- ggplot(mt, aes(x = mpg, y = wt, label = car)) +
     geom_point() +
     theme_cie()

# A chart with CIE labels
p +
    cie_label()

# A chart with CIE labels that are repelled from each other
p +
    cie_label_repel()


}
