% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cie_colours.R
\docType{data}
\name{cie_lightyellow}
\alias{cie_lightyellow}
\title{cie_lightyellow}
\format{
An object of class \code{character} of length 1.
}
\usage{
cie_lightyellow
}
\description{
cie_lightyellow
}
\keyword{datasets}
