% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cie_pal.R
\name{cie_pal}
\alias{cie_pal}
\title{Create a CIE-appropriate palette for your chart.}
\usage{
cie_pal(n = 0, reverse = FALSE, faded_level = 0, faded = FALSE)
}
\arguments{
\item{n}{Numeric. The number of levels in your colour scale. Minimum value is
  1, maximum is 10. Using more than 6 is not recommended. If you don't
  specify `n`, a five-colour palette will be used, which may not look right.
  Specify `n`.

  By default, n = 2 will give you light orange and dark orange. Use n = "2a"
  if you want light orange and red.}

\item{reverse}{Logical. FALSE by default. Setting to TRUE reverses the
standard colour order. Standard colour order runs from light to dark. If
you set reverse to TRUE, colours will run from dark to light.}

\item{faded_level}{Integer between 0 (no fade; the default) and 8 (most faded).}

\item{faded}{Deprecated - use \code{faded_level}. Logical. FALSE by default. Setting to TRUE returns the faded
variations of the standard colours. If TRUE and \code{faded_level = 0}, sets \code{faded_level} to 5.}
}
\description{
Create a CIE-appropriate palette for your chart.
}
\examples{
library(ggplot2)

p <- ggplot(mtcars, aes(x = wt, y = mpg, col = factor(cyl))) +
    geom_point() +
    theme_cie() +
    scale_colour_manual(values = cie_pal(n = 3))

p

# Alternatively, use cie_colour_manual(), which is a wrapper
# around scale_colour_manual():

p <- ggplot(mtcars, aes(x = wt, y = mpg, col = factor(cyl))) +
    geom_point() +
    theme_cie() +
    cie_colour_manual(n = 3)

p

}
