% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cie_colours.R
\docType{data}
\name{cie_palette_set}
\alias{cie_palette_set}
\title{A list of CIE colour sets}
\format{
An object of class \code{list} of length 6.
}
\usage{
cie_palette_set
}
\description{
"full": red, dark orange, light orange, yellow, light yellow
"full_f": faded version of "full"
"light": light orange, yellow, light yellow
"dark": red, dark orange, light orange
"diverging": red, faded red, white, faded light orange, light orange
"grey": grey 1, grey 2, grey 3, grey 4, grey 5
}
\keyword{datasets}
