% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cie_pal.R
\docType{data}
\name{cie_palettes}
\alias{cie_palettes}
\title{Predefined cie colours combined into palettes}
\format{
An object of class \code{list} of length 3.
}
\usage{
cie_palettes
}
\description{
This is a list of cie colours combined into palettes. The palettes are used
for different plots and maps.
}
\keyword{datasets}
