% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cie_scales.R
\name{cie_scale}
\alias{cie_colour_manual}
\alias{cie_fill_manual}
\title{Convenient functions to set CIE-appropriate palettes}
\usage{
cie_colour_manual(
  n = 0,
  reverse = FALSE,
  discrete = TRUE,
  faded_level = 0,
  faded = FALSE,
  palette = "full",
  ...
)

cie_fill_manual(
  n = 0,
  reverse = FALSE,
  discrete = TRUE,
  faded_level = 0,
  faded = FALSE,
  palette = "full",
  ...
)
}
\arguments{
\item{n}{Numeric. The number of levels in your colour scale. Minimum value is
1, maximum is 7. Passed to \code{cie_pal}; see ?cie_pal for more
information.}

\item{reverse}{Logical. FALSE by default. Setting to TRUE reverses the
standard colour order.}

\item{discrete}{Logical. TRUE by default. Setting to FALSE generates a
continuous colour scale.}

\item{faded_level}{Integer between 0 (no fade; the default) and 8 (most faded).}

\item{faded}{Deprecated - use \code{faded_level}. Logical. FALSE by default. Setting to TRUE returns the faded
variations of the standard colours. If TRUE and \code{faded_level = 0}, sets \code{faded_level} to 5.}

\item{palette}{Sets the colours that will form the continuous palette when
  discrete = FALSE. One of:

\describe{
  \item{full}{The default. Red, dark orange, light orange, yellow, light
yellow}
  \item{full_f}{ faded version of "full"}
  \item{light}{ light
orange, yellow, light yellow}
  \item{dark}{ red, dark orange, light orange}
  \item{diverging}{ red, faded red, white, faded light orange, light orange}
  \item{grey}{ grey 1, grey 2, grey 3, grey 4, grey 5}
}}

\item{...}{arguments passed to ggplot2 scales}
}
\description{
Convenient functions to set CIE-appropriate palettes
}
\examples{

library(ggplot2)

ggplot(data = mtcars, aes(x = wt, y = mpg, col = factor(cyl))) +
   geom_point() +
   cie_colour_manual(n = 3) +
   theme_cie()

}
