% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cie_save_pptx.R
\name{create_pptx_shell}
\alias{create_pptx_shell}
\title{Create a Powerpoint document containing no content other than the
template + slide-specific speaker notes (extracted from a ggplot2 object)}
\usage{
create_pptx_shell(p, filename, type)
}
\arguments{
\item{p}{list containing ggplot2 object(s)}

\item{filename}{filename (incl. path if needed) to save a .pptx file}

\item{type}{graph type}
}
\description{
Create a Powerpoint document containing no content other than the
template + slide-specific speaker notes (extracted from a ggplot2 object)
}
\keyword{internal}
