% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cie_save_pptx.R
\name{create_slide_shell}
\alias{create_slide_shell}
\title{Generate the R Markdown code for an individual slide in a PPTX shell}
\usage{
create_slide_shell(p, filename)
}
\description{
Generate the R Markdown code for an individual slide in a PPTX shell
}
\keyword{internal}
