% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_default_aesthetics.R
\name{.prep_aes}
\alias{.prep_aes}
\title{Prepare \code{ggplot2} geom_ defaults}
\usage{
.prep_aes(geom, aes, setting)
}
\arguments{
\item{geom}{(character) the geom to change (e.g. \code{point} or
\code{line})}

\item{aes}{(character) the aesthetic to change (e.g. \code{colour} or
\code{alpha})}

\item{setting}{(various) what to set the aesthetic (e.g. a colour or a
number)}
}
\value{
the default aesthetic ready to be set
}
\description{
ggplot2 geoms have default aesthetics which can be changed for each
session. This function is a generalised way to set geom defaults. It only
changes settings where the default is not NA; no need for \code{geom_point}
to have a default font family.
}
\details{
This is used inside `.set_cie_aesthetics()`
}
