% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_default_aesthetics.R
\name{.set_cie_aesthetics}
\alias{.set_cie_aesthetics}
\title{Set the default ggplot2 aesthetics to cie branding}
\usage{
.set_cie_aesthetics()
}
\description{
A wrapper for a bunch of other functions to set ggplot2 default aesthetics.
}
