% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/font_utils.R
\name{get_cie_font}
\alias{get_cie_font}
\title{Get font for chart type}
\usage{
get_cie_font(type = c("normal", "slide"), element = c("body", "title"))
}
\arguments{
\item{type}{Either "normal" or "slide"}

\item{element}{For "slide" type, either "title" or "body". Ignored for "normal".}
}
\value{
Character string with font family name
}
\description{
Get font for chart type
}
\keyword{internal}
