% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overleaf_utils.R
\name{get_overleaf_base_path}
\alias{get_overleaf_base_path}
\title{Get the Overleaf base path}
\usage{
get_overleaf_base_path()
}
\value{
Character string containing the path to the Apps/Overleaf folder
}
\description{
Finds the Apps/Overleaf folder in the user's Centre for International Economics
  Dropbox by searching one level down from the Dropbox root, excluding
  'data' and 'CIE Team' folders.
}
\keyword{internal}
