% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cie_pal.R
\name{make_cie_pal_discrete}
\alias{make_cie_pal_discrete}
\title{Create a cie colour palette}
\usage{
make_cie_pal_discrete(n)
}
\arguments{
\item{n}{how many colours to return}
}
\description{
This function takes a the cie graph colour palette and returns a vector of colours equal to n.
It is used in \code{\link{scale_colour_cie}} and \code{\link{scale_fill_cie}} to make the discrete
colour scale as the order of colours is specific in the cie branding guides and so using an interpolated scale
does not work.
}
\seealso{
\code{\link{cie_palettes}}
}
