% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cie_pal.R
\docType{data}
\name{palette_order}
\alias{palette_order}
\title{Set the palette order as needed in the}
\format{
An object of class \code{list} of length 2.
}
\usage{
palette_order
}
\description{
This is a list of cie colours combined into palettes. The palettes are used
for different plots and maps.
}
\keyword{datasets}
