% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cie_save_pptx.R
\name{return_script_and_chart_location}
\alias{return_script_and_chart_location}
\title{Return a charter vector of the filepath of the script and powerpoint location}
\usage{
return_script_and_chart_location(filename)
}
\arguments{
\item{filename}{the filename of the powerpoint location}
}
\value{
a character
}
\description{
An internal function used in `add_graph_to_pptx` to add the location of the script that
made a chart and the location where the chart is saved.
}
