% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cie_scales.R
\name{scale_colour_cie}
\alias{scale_colour_cie}
\alias{scale_fill_cie}
\title{Scale constructor for cie colours}
\usage{
scale_colour_cie(palette = NULL, discrete = FALSE, reverse = FALSE, ...)

scale_fill_cie(palette = NULL, discrete = FALSE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in \code{cie_palettes}}

\item{discrete}{Boolean indicating whether colour aesthetic is discrete or
not}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to \code{\link[ggplot2]{discrete_scale}} or
\code{\link[ggplot2]{scale_color_gradientn}}, used respectively when discrete is
TRUE or FALSE}
}
\description{
If you get an error about not enough colours, try a different palette (such
as \code{graph}) or set \code{discrete = FALSE}. If you are plotting discrete colours
and need more than 10 colours, then you will have to make your own palette using
\code{\link{make_cie_pal}}.
}
\section{Functions}{
\itemize{
\item \code{scale_colour_cie()}: For colour scales

\item \code{scale_fill_cie()}: For fill scales

}}
